src/requirements.txt
-i https://pypi.python.org/simple
asn1crypto==0.24.0
attrs==19.1.0
aws-encryption-sdk==1.4.0
boto3==1.9.158
botocore==1.12.158
cffi==1.12.3
cryptography==2.6.1
docutils==0.14
jmespath==0.9.4
jsonschema==3.0.1
pycparser==2.19
pyrsistent==0.15.2
python-dateutil==2.8.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.25.3 ; python_version >= '3.4'
wrapt==1.11.1
